fix: fixing some DTOs and Responses to improve in front-end

+++ API_JAVA/deutschland-game/src/main/java/br/com/etechoracio/deutschland_game/dtos/LoadConsequenciasConquistasDto.java
@@ -1,4 +1,4 @@
 package br.com.etechoracio.deutschland_game.dtos;
 
-public record LoadConsequenciasConquistasDto(Long id_conquistas, String nome_conquistas, Double valor_acrescentado) {
+public record LoadConsequenciasConquistasDto(Long id_conquistas, String nome_conquistas, Integer valor_acrescentado) {
 }
